refactor: unify redux stores
