+++ packages/frontend/src/components/DataViz/store/selectors.ts
@@ -4,7 +4,7 @@ import {
     type PivotChartLayout,
 } from '@lightdash/common';
 import { createSelector } from 'reselect';
-import { type RootState } from '.';
+import { type RootState } from '../../../features/sqlRunner/store';
 import { type TableVizState } from './tableVisSlice';
 
 const selectBarChartConfigState = (
