+++ packages/frontend/src/features/semanticViewer/components/SemanticViewerVizConfig/TableVizConfig.tsx
@@ -5,14 +5,14 @@ import { type FC } from 'react';
 import MantineIcon from '../../../../components/common/MantineIcon';
 import { TableFieldIcon } from '../../../../components/DataViz/Icons';
 import {
-    useVizDispatch,
-    useVizSelector,
-} from '../../../../components/DataViz/store';
-import {
     updateColumnVisibility,
     updateFieldLabel,
 } from '../../../../components/DataViz/store/tableVisSlice';
 import { Config } from '../../../../components/VisualizationConfigs/common/Config';
+import {
+    useAppDispatch as useVizDispatch,
+    useAppSelector as useVizSelector,
+} from '../../../../features/sqlRunner/store/hooks';
 
 const TableVisConfiguration: FC<{ columns: VizColumn[] }> = ({ columns }) => {
     const dispatch = useVizDispatch();
