+++ packages/frontend/src/features/semanticViewer/components/SidebarViews.tsx
@@ -2,8 +2,8 @@ import { Center, Loader, NavLink, Stack, Text } from '@mantine/core';
 import { IconTable } from '@tabler/icons-react';
 import MantineIcon from '../../../components/common/MantineIcon';
 import SuboptimalState from '../../../components/common/SuboptimalState/SuboptimalState';
+import { useAppDispatch, useAppSelector } from '../../sqlRunner/store/hooks';
 import { useSemanticLayerViews } from '../api/hooks';
-import { useAppDispatch, useAppSelector } from '../store/hooks';
 import { selectSemanticLayerInfo } from '../store/selectors';
 import { enterView } from '../store/semanticViewerSlice';
